Move slice index computation logic into a separate mod
We have too much code in `values/mod.rs`.
