+++ starlark/src/values/list.rs
@@ -16,6 +16,7 @@
 use crate::stdlib::list::LIST_REMOVE_ELEMENT_NOT_FOUND_ERROR_CODE;
 use crate::values::error::{RuntimeError, ValueError};
 use crate::values::iter::TypedIterable;
+use crate::values::slice_indices::convert_slice_indices;
 use crate::values::*;
 use std::cmp::Ordering;
 use std::fmt;
@@ -203,8 +204,7 @@ impl TypedValue for List {
         stop: Option<Value>,
         stride: Option<Value>,
     ) -> ValueResult {
-        let (start, stop, stride) =
-            Value::convert_slice_indices(self.length()?, start, stop, stride)?;
+        let (start, stop, stride) = convert_slice_indices(self.length()?, start, stop, stride)?;
         Ok(Value::from(tuple::slice_vector(
             start,
             stop,
