+++ starlark/src/values/mod.rs
@@ -1201,147 +1201,6 @@ impl PartialEq for Value {
 }
 impl Eq for Value {}
 
-impl dyn TypedValueDyn {
-    // To be calleds by convert_slice_indices only
-    fn convert_index_aux(
-        len: i64,
-        v1: Option<Value>,
-        default: i64,
-        min: i64,
-        max: i64,
-    ) -> Result<i64, ValueError> {
-        if let Some(v) = v1 {
-            if v.get_type() == "NoneType" {
-                Ok(default)
-            } else {
-                match v.to_int() {
-                    Ok(x) => {
-                        let i = if x < 0 { len + x } else { x };
-                        if i < min {
-                            Ok(min)
-                        } else if i > max {
-                            Ok(max)
-                        } else {
-                            Ok(i)
-                        }
-                    }
-                    Err(..) => Err(ValueError::IncorrectParameterType),
-                }
-            }
-        } else {
-            Ok(default)
-        }
-    }
-
-    /// Macro to parse the index for at/set_at methods.
-    ///
-    /// Return an `i64` from self corresponding to the index recenterd between 0 and len.
-    /// Raise the correct errors if the value is not numeric or the index is out of bound.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use starlark::values::*;
-    /// # assert!(
-    /// Value::new(6).convert_index(7).unwrap() == 6
-    /// # );
-    /// # assert!(
-    /// Value::new(-1).convert_index(7).unwrap() == 6
-    /// # );
-    /// ```
-    ///
-    /// The following examples would return an error
-    /// ```rust
-    /// # use starlark::values::*;
-    /// # use starlark::values::error::*;
-    /// # use starlark::values::string;
-    /// # assert!(
-    /// Value::from("a").convert_index(7) == Err(ValueError::IncorrectParameterType)
-    /// # );
-    /// # assert!(
-    /// Value::new(8).convert_index(7) == Err(ValueError::IndexOutOfBound(8))   // 8 > 7 = len
-    /// # );
-    /// # assert!(
-    /// Value::new(-8).convert_index(7) == Err(ValueError::IndexOutOfBound(-1)) // -8 + 7 = -1 < 0
-    /// # );
-    /// ```
-    pub fn convert_index(&self, len: i64) -> Result<i64, ValueError> {
-        match self.to_int_dyn() {
-            Ok(x) => {
-                let i = if x < 0 {
-                    len.checked_add(x).ok_or(ValueError::IntegerOverflow)?
-                } else {
-                    x
-                };
-                if i < 0 || i >= len {
-                    Err(ValueError::IndexOutOfBound(i))
-                } else {
-                    Ok(i)
-                }
-            }
-            Err(..) => Err(ValueError::IncorrectParameterType),
-        }
-    }
-}
-
-impl Value {
-    /// Parse indices for slicing.
-    ///
-    /// Takes the object length and 3 optional values and returns `(i64, i64, i64)`
-    /// with those index correctly converted in range of length.
-    /// Return the correct errors if the values are not numeric or the stride is 0.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use starlark::values::*;
-    /// let six      = Some(Value::new(6));
-    /// let minusone = Some(Value::new(-1));
-    /// let ten      = Some(Value::new(10));
-    ///
-    /// # assert!(
-    /// Value::convert_slice_indices(7, six, None, None).unwrap() == (6, 7, 1)
-    /// # );
-    /// # assert!(
-    /// Value::convert_slice_indices(7, minusone.clone(), None, minusone.clone()).unwrap()
-    ///        == (6, -1, -1)
-    /// # );
-    /// # assert!(
-    /// Value::convert_slice_indices(7, minusone, ten, None).unwrap() == (6, 7, 1)
-    /// # );
-    /// ```
-    pub fn convert_slice_indices(
-        len: i64,
-        start: Option<Value>,
-        stop: Option<Value>,
-        stride: Option<Value>,
-    ) -> Result<(i64, i64, i64), ValueError> {
-        let stride = stride.unwrap_or_else(|| Value::new(1));
-        let stride = if stride.get_type() == "NoneType" {
-            Ok(1)
-        } else {
-            stride.to_int()
-        };
-        match stride {
-            Ok(0) => Err(ValueError::IndexOutOfBound(0)),
-            Ok(stride) => {
-                let def_start = if stride < 0 { len - 1 } else { 0 };
-                let def_end = if stride < 0 { -1 } else { len };
-                let clamp = if stride < 0 { -1 } else { 0 };
-                let start =
-                    TypedValueDyn::convert_index_aux(len, start, def_start, clamp, len + clamp);
-                let stop = TypedValueDyn::convert_index_aux(len, stop, def_end, clamp, len + clamp);
-                match (start, stop) {
-                    (Ok(s1), Ok(s2)) => Ok((s1, s2, stride)),
-                    (Err(x), ..) => Err(x),
-                    (Ok(..), Err(x)) => Err(x),
-                }
-            }
-            _ => Err(ValueError::IncorrectParameterType),
-        }
-    }
-}
-
 impl Value {
     /// Get a reference to underlying data or `None`
     /// if contained object has different type than requested.
@@ -1379,10 +1238,6 @@ impl Value {
             None
         })
     }
-
-    pub fn convert_index(&self, len: i64) -> Result<i64, ValueError> {
-        self.value_holder().convert_index(len)
-    }
 }
 
 // Submodules
@@ -1397,6 +1252,7 @@ pub mod iter;
 pub mod list;
 pub mod none;
 pub mod range;
+pub mod slice_indices;
 pub mod string;
 pub mod tuple;
 
@@ -1412,37 +1268,6 @@ use std::any::Any;
 mod tests {
     use super::*;
     use std::iter;
-
-    #[test]
-    fn test_convert_index() {
-        assert_eq!(Ok(6), Value::new(6).convert_index(7));
-        assert_eq!(Ok(6), Value::new(-1).convert_index(7));
-        assert_eq!(
-            Ok((6, 7, 1)),
-            Value::convert_slice_indices(7, Some(Value::new(6)), None, None)
-        );
-        assert_eq!(
-            Ok((6, -1, -1)),
-            Value::convert_slice_indices(7, Some(Value::new(-1)), None, Some(Value::new(-1)))
-        );
-        assert_eq!(
-            Ok((6, 7, 1)),
-            Value::convert_slice_indices(7, Some(Value::new(-1)), Some(Value::new(10)), None)
-        );
-        // Errors
-        assert_eq!(
-            Err(ValueError::IncorrectParameterType),
-            Value::from("a").convert_index(7)
-        );
-        assert_eq!(
-            Err(ValueError::IndexOutOfBound(8)),
-            Value::new(8).convert_index(7)
-        );
-        assert_eq!(
-            Err(ValueError::IndexOutOfBound(-1)),
-            Value::new(-8).convert_index(7)
-        );
-    }
 
     #[test]
     fn can_implement_compare() {
